###Question 1###
While I deeply appreciate the fact that programming makes the world a
better and more efficient place, coding has special draw for me because
I find the actual process to be extremely satisfying. It aligns with
things I already enjoy: it requires a high attention to detail and is
based on problem solving; it is project oriented; and keeps me wanting
to learn more. I have pursued careers in management, physical fitness,
and patient care and have never been as engaged as I am when I am
solving algorithms.

The following is a list of classes I’ve taken to pursue additional
knowledge in coding, and includes online and on-site experiences.

_General Assembly_
(1) Coding for Beginners: HTML and CSS
(2) Dash - HTML, CSS and Javascript

_Free Code Camp_
25 hours Front Stack Development Map

_Skillcrush_
HTML and CSS Blueprint Class

_Coding Dojo_
Completed algorithm app

###Question 2###

For the next two years, my focus will be on laying a foundation that
prepares me to learn and build confidence in the field of software
development. I hope to have spent a year at Ada, and to have one year’s
experience as a junior developer.

The next five years are harder for me to predict. I am curious about
where my passion in this field will ultimately lie, and look forward to
exploring new areas I never knew I would be excited about. I do know
that I’m interested in extending my experience in software development
to creating apps that help users conserve resources and monitor how
their actions affect our planet.

An area that’s of particular importance to me is water conservation.
The cost of water is cheap, and as a result, there is little to no
incentive to conserve it. Additionally, users have to take the extra
step of physically checking their water meter if they want to know how
much they’re using. I am interested in creating an app that works with
current water meter technology to allow users to monitor usage from
their mobile phone.

###Question 3###

You know that little voice in your head that lets you know when you’re
doing something you shouldn’t be doing? I heard it and I ignored it.

Several years ago, because of a general interest in fitness and the
workings of the human body, I decided a career in physical therapy made
sense for me. Being an all-or-nothing type of person, I went back to
school and finished the necessary prerequisites and gained over 500
hours experience volunteering in the field. But amidst all of this
drive I had for my new-found career, there was that voice telling me it
wasn’t quite right. I loved the knowledge necessary for the job, but I
didn’t love patient care. I didn’t love the idea of having to be in an
enclosed space and make small talk for 9 hours a day, or having to
motivate patients who weren’t doing their exercises at home. I didn’t
acknowledge that the voice was speaking to me until I had fulfilled all
of the requirements to take the final step of entering a graduate
school program. If I had listened to the voice years earlier, I could
have started down another career path--like coding--that didn’t give me
any doubts.

At this point, I know myself and my strengths and weaknesses. As I type
the answers to these essay questions, there is no little voice--only
the peace that I am right where I should be.

###Question 4###

As my senior year of high school came to a close, the four seniors in
my orchestra decided to play as a quartet for the graduation ceremony.
String quartets traditionally consist of two violins, a viola and a
cello. We were four violinists. One person who’d played viola before
filled in that position, but we’d chosen to play Green Day’s
cheesy-but-nostalgic “Good Riddance,” which required the addition of a
cellist. I’d played the violin for 10 years: reading treble clef and
handling my violin was second nature. But, I’d always found the cello
to be the sexiest stringed instrument, so I volunteered to learn for
the ceremony.

I borrowed a public school-grade cello and took it home to try to
figure out what I was doing. Fortunately, a cellist friend came over to
give me a crash course in reading bass clef and how to play the
instrument. The size upgrade was awkward but doable; the most
challenging part was learning to read the music. I had to translate
each note from bass to treble using a system of musical thirds in my
head. When it was time to play, I was nowhere near fluent in bass clef,
but using my simple system, I was able to play this new instrument with
old friends at my high school graduation--something I’ll never forget.